Reorganize the file structure
